Title: 5와 7의 배수를 제외하기, Time: 151ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240311/5와 7의 배수를 제외하기/excluding-multiples-of-5-and-7.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int cnt = 0, sum = 0;
+        double cnt = 0, sum = 0;
         for(int i = 0 ; i < n ; i++){
             int num = sc.nextInt();
             if(num % 5 == 0 || num % 7 == 0){
@@ -13,7 +13,7 @@ public class Main {
             sum += num;
         }
         double avg = sum / cnt;
-        System.out.println(sum);
+        System.out.println((int)sum);
         System.out.printf("%.1f", avg);
     }
 }
